refactor: todoList에 아이템을 추가할 때 localStorage에 저장 되도록 변경
