+++ src/todoList/countTodoItem.js
@@ -1,6 +1,5 @@
 export const countTodoItem = () => {
   const $todoCount = document.querySelector('.todo-count > strong');
-
   const $todoList = document.querySelectorAll('.todo-list > li');
   const $todoListHidden = document.querySelectorAll(
     '.todo-list > li[class="hidden"]',
