+++ src/todoList/render.js
@@ -1,0 +1,6 @@
+export const render = () => {
+  const todos = JSON.parse(localStorage.getItem('todos'));
+  const $todoList = document.querySelector('.todo-list');
+
+  $todoList.innerHTML = todos;
+};
